# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What Deck API",
        "description": "Magic: The Gathering card collection management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/cards": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "List all cards",
                "description": "Get a paginated list of Magic: The Gathering cards with optional filtering",
                "operationId": "e1710b39810207a38a8d9fa5d8ed4407",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter cards by type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Instant"
                        }
                    },
                    {
                        "name": "subtype",
                        "in": "query",
                        "description": "Filter cards by subtype",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Spell"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search cards by title",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Lightning"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Card"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cards"
                ],
                "summary": "Create a new card",
                "description": "Create a new Magic: The Gathering card",
                "operationId": "f2ac6dc4dc4f0cc90f016cbe8b02fb10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Lightning Bolt"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/cards/lightning-bolt.jpg",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lightning Bolt deals 3 damage to any target.",
                                        "nullable": true
                                    },
                                    "cost": {
                                        "type": "string",
                                        "example": "{R}",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Instant"
                                    },
                                    "subtype": {
                                        "type": "string",
                                        "example": "Spell",
                                        "nullable": true
                                    },
                                    "power": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": null,
                                        "nullable": true
                                    },
                                    "toughness": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cards/{id}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Get a specific card",
                "description": "Retrieve a specific Magic: The Gathering card with its relationships",
                "operationId": "af26c5e68f333f1dfadd1bfa227eb617",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cards"
                ],
                "summary": "Update a card",
                "description": "Update an existing Magic: The Gathering card",
                "operationId": "fe7b2d31eb73d88c9c7884acaee234d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Lightning Bolt"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/cards/lightning-bolt.jpg",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lightning Bolt deals 3 damage to any target.",
                                        "nullable": true
                                    },
                                    "cost": {
                                        "type": "string",
                                        "example": "{R}",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Instant"
                                    },
                                    "subtype": {
                                        "type": "string",
                                        "example": "Spell",
                                        "nullable": true
                                    },
                                    "power": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": null,
                                        "nullable": true
                                    },
                                    "toughness": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Card updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cards"
                ],
                "summary": "Delete a card",
                "description": "Delete a Magic: The Gathering card (only if no instances exist)",
                "operationId": "bdb2c37c481aff6ee4b63a61add8471d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot delete card with existing instances",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Paginated Response",
                "description": "Standard paginated response format",
                "properties": {
                    "data": {
                        "description": "Array of resource items",
                        "type": "array",
                        "items": {}
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 67
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=5"
                            },
                            "prev": {
                                "type": "string",
                                "format": "url",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Validation Error Response",
                "description": "Response format for validation errors",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ],
                            "type": [
                                "The type field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "card": [
                                "This card has existing instances and cannot be deleted"
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Card": {
                "title": "Card Resource",
                "description": "Magic: The Gathering card resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Lightning Bolt"
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/cards/lightning-bolt.jpg",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Lightning Bolt deals 3 damage to any target.",
                        "nullable": true
                    },
                    "cost": {
                        "type": "string",
                        "example": "{R}",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "Instant"
                    },
                    "subtype": {
                        "type": "string",
                        "example": "Spell",
                        "nullable": true
                    },
                    "power": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "toughness": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "edition": {
                        "type": "string",
                        "example": "dft",
                        "nullable": true
                    },
                    "collector_number": {
                        "type": "string",
                        "example": "187",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "card_instances": {
                        "description": "Card instances (only when relationship is loaded)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CardInstance"
                        }
                    },
                    "collections": {
                        "description": "Collections containing this card (only when relationship is loaded)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Collection"
                        }
                    },
                    "decks": {
                        "description": "Decks containing this card (only when relationship is loaded)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Deck"
                        }
                    },
                    "total_instances": {
                        "description": "Total number of instances of this card",
                        "type": "integer",
                        "example": 5
                    },
                    "instances_in_decks": {
                        "description": "Number of instances currently in decks",
                        "type": "integer",
                        "example": 2
                    },
                    "available_instances": {
                        "description": "Number of instances not in decks",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}